Uto 2.0 程式碼總攬

+++ cmds/main.py
@@ -1,16 +1,15 @@
 # ping information
 import discord
-from discord import message
-from discord import channel
 from discord import embeds
 from discord.ext import commands
 from discord.ext.commands import errors
+from discord.ext.commands.core import dm_only
 from core.classes import Cog_Extension
-import json, asyncio, datetime
+import json, asyncio
 from fake_useragent import UserAgent
 import requests
-from bs4 import BeautifulSoup
-import sys
+import datetime
+from datetime import timedelta
 
 
 with open('setting.json','r',encoding='utf8') as jfile:
@@ -27,14 +26,15 @@ class Main(Cog_Extension):
     async def ping(self,ctx):
         await ctx.send(f"{round(self.bot.latency*1000)}(ms)")
 
-#    @commands.command()
-#    async def settime(self,ctx,time):
-#        self.counter = 0
-#        with open('time.json','r',encoding='utf8') as jfile:
-#            jdat = json.load(jfile)
-#        jdat['time'] = time
-#        with open('time.json','w',encoding='utf8') as jfile:
-#            json.dump(jdat,jfile,indent=4)
+    
+    @commands.command(aliases=['time','dly'])
+    async def delay(self,ctx,msg):
+        try:
+            msgid = await ctx.send(f'開始計時 {msg} 秒！')
+            await asyncio.sleep(int(msg))
+            await ctx.send('時間到！！')
+        except:
+            await msgid.edit(content = '❌ | 發生錯誤')
     @commands.command(aliases=['dcr'])
     async def dcard(self,ctx,msg):
         req = ""
@@ -79,13 +79,13 @@ class Main(Cog_Extension):
         msg = await ctx.send(embed=embed)
         await msg.add_reaction('⏩')
 
-    @commands.command(aliases=['helps'])
+    @commands.command(aliases=['helps','cmd','cmds'])
     async def command(self,ctx):
         text =""
         bn = int(command['count'])
         an = int(command['nowpage'])
         text += str(f"{command[f'{an}']}")
-        embed=discord.Embed(title=f"~ :book: 指 令 說 明 command information({an}/{bn} ~)", description=f":level_slider:指令 <必要函數> (非必要函數)\n\n{text}", color=0x00ff64, timestamp=ctx.message.created_at)
+        embed=discord.Embed(title=f"~ :book: 指 令 說 明 command information({an}/{bn})~", description=f":level_slider:指令 <必要函數> (非必要函數)\n{text}", color=0x00ff64, timestamp=ctx.message.created_at)
         embed.set_footer(text="discord bot name : Uto , 顯示時間:")
         msg = await ctx.send(embed=embed)
         await msg.add_reaction('⬅️')
@@ -99,32 +99,34 @@ class Main(Cog_Extension):
     @commands.command(aliases=['ts'])
     async def test(self,ctx,*,msg):
         await ctx.message.delete()
-        embed=discord.Embed(title="📜考試", description=f"{msg}", color=0xa686fe, timestamp=ctx.message.created_at)
-        embed.set_footer(text=f"發起人 {ctx.author}",icon_url=ctx.author.avatar_url)
+        text = msg
+        embed=discord.Embed(title="📜考試", description=f"{text}", color=0xa686fe, timestamp=ctx.message.created_at)
         embed.set_footer(text=f"發起人 {ctx.author},      發起時間:",icon_url=ctx.author.avatar_url)
-        msg = await ctx.send(embed=embed)
-        await msg.add_reaction('✅')
+        msgid = await ctx.send(embed=embed)
+        await msgid.add_reaction('✅')
 
 
     @commands.command(aliases=['wk'])
     async def work(self,ctx,*,msg):
         await ctx.message.delete()
-        embed=discord.Embed(title="📚作業", description=f"{msg}", color=0xa686fe, timestamp=ctx.message.created_at)
+        text = msg
+        embed=discord.Embed(title="📚作業", description=f"{text}", color=0xa686fe, timestamp=ctx.message.created_at)
         embed.set_footer(text=f"發起人 {ctx.author},      發起時間:",icon_url=ctx.author.avatar_url)
-        msg = await ctx.send(embed=embed)
-        await msg.add_reaction('✅')
+        msgid = await ctx.send(embed=embed)
+        await msgid.add_reaction('✅')
+
 
     @commands.command(aliases=['vt'])
     async def vote(self,ctx,*,msg):
         await ctx.message.delete()
-        await ctx.send(f"{ctx.message.guild.default_role}")
+        msg2id = await ctx.send(f"{ctx.message.guild.default_role}")
+        text = msg
         embed=discord.Embed(title="📋投票", color=0x1f7b1e, timestamp=ctx.message.created_at)
-        embed.add_field(name="內容", value=f"{msg}")
-        embed.set_footer(text=f"發起人 {ctx.author}",icon_url=ctx.author.avatar_url)
+        embed.add_field(name="內容", value=f"{text}")
         embed.set_footer(text=f"發起人 {ctx.author},      發起時間:",icon_url=ctx.author.avatar_url)
-        msg = await ctx.send(embed=embed)
-        await msg.add_reaction('✅')
-        await msg.add_reaction('❌')
+        msgid = await ctx.send(embed=embed)
+        await msgid.add_reaction('✅')
+        await msgid.add_reaction('❌')
 
 
     @commands.command(aliases=['wn'])
@@ -133,15 +135,22 @@ class Main(Cog_Extension):
         member = ctx.guild.get_member(int(id))
         print(member)
         embed=discord.Embed(title="⚠️警告", description=f"懲處🚫：\n發現成員 `{member}` 違反規定\n原因：{msg2}\nUto判定懲處：{msg}", color=0xa686fe, timestamp=ctx.message.created_at)
-        embed.set_footer(text=f"授予審理人 {ctx.author}",icon_url=ctx.author.avatar_url)
         embed.set_footer(text=f"授予審理人 {ctx.author},      發起時間:",icon_url=ctx.author.avatar_url)
         msg = await ctx.send(embed=embed)
 
 
     @commands.command(aliases=['s'])
     async def say(self,ctx,*,msg):
-        await ctx.message.delete()
-        msg = await ctx.send(msg)
+        autor = self.bot.get_user(561731559493861398)
+        if(autor==ctx.author):
+            try:
+                await ctx.message.delete()
+            except:
+                pass
+            msg = await ctx.send(msg)
+        else:
+            await ctx.send("❌| 你沒有權限使用此指令")
+        
 
 
     @commands.command(aliases=['clean','purge'])
@@ -152,6 +161,21 @@ class Main(Cog_Extension):
             await ctx.channel.purge(limit=int(num))
             await ctx.send(f"✅| 成功清除{num}則訊息")
 
+    @commands.command(aliases=['temp'])
+    async def tp(self,ctx):
+        autor = self.bot.get_user(561731559493861398)
+        if(autor==ctx.author):
+            await ctx.send(f"getit！It's you！{autor}")
+        else:
+            await ctx.send(f'{autor} {ctx.author}')
+
+
+
+    @commands.command(aliases=['nickurl','url'])
+    async def nick(self,ctx,msg,url):
+        await ctx.message.delete()
+        embed=discord.Embed(description=f"[{msg}]({url})")
+        msg = await ctx.send(embed=embed)
     
 def setup(bot):
     bot.add_cog(Main(bot))
